+++ packages/go_router/lib/src/parser.dart
@@ -8,11 +8,24 @@ import 'dart:math';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 
-import '../go_router.dart';
 import 'configuration.dart';
 import 'information_provider.dart';
 import 'logging.dart';
 import 'match.dart';
+import 'route.dart';
+import 'router.dart';
+
+/// The function signature of [GoRouteInformationParser.onParserException].
+///
+/// The `routeMatchList` parameter contains the exception explains the issue
+/// occurred.
+///
+/// The returned [RouteMatchList] is used as parsed result for the
+/// [GoRouterDelegate].
+typedef ParserExceptionHandler = RouteMatchList Function(
+  BuildContext context,
+  RouteMatchList routeMatchList,
+);
 
 /// Converts between incoming URLs and a [RouteMatchList] using [RouteMatcher].
 /// Also performs redirection using [RouteRedirector].
@@ -20,10 +33,17 @@ class GoRouteInformationParser extends RouteInformationParser<RouteMatchList> {
   /// Creates a [GoRouteInformationParser].
   GoRouteInformationParser({
     required this.configuration,
+    required this.onParserException,
   }) : _routeMatchListCodec = RouteMatchListCodec(configuration);
 
   /// The route configuration used for parsing [RouteInformation]s.
   final RouteConfiguration configuration;
+
+  /// The exception handler that is called when parser can't handle the incoming
+  /// uri.
+  ///
+  /// This method must return a [RouteMatchList] for the parsed result.
+  final ParserExceptionHandler? onParserException;
 
   final RouteMatchListCodec _routeMatchListCodec;
 
@@ -50,7 +70,13 @@ class GoRouteInformationParser extends RouteInformationParser<RouteMatchList> {
       // the state.
       final RouteMatchList matchList =
           _routeMatchListCodec.decode(state as Map<Object?, Object?>);
-      return debugParserFuture = _redirect(context, matchList);
+      return debugParserFuture = _redirect(context, matchList)
+          .then<RouteMatchList>((RouteMatchList value) {
+        if (value.isError && onParserException != null) {
+          return onParserException!(context, value);
+        }
+        return value;
+      });
     }
 
     late final RouteMatchList initialMatches;
@@ -70,6 +96,17 @@ class GoRouteInformationParser extends RouteInformationParser<RouteMatchList> {
       context,
       initialMatches,
     ).then<RouteMatchList>((RouteMatchList matchList) {
+      if (matchList.isError && onParserException != null) {
+        return onParserException!(context, matchList);
+      }
+
+      assert(() {
+        if (matchList.isNotEmpty) {
+          assert(!(matchList.last.route as GoRoute).redirectOnly,
+              'A redirect-only route must redirect to location different from itself.\n The offending route: ${matchList.last.route}');
+        }
+        return true;
+      }());
       return _updateRouteMatchList(
         matchList,
         baseRouteMatchList: state.baseRouteMatchList,