+++ packages/go_router/test/extension_test.dart
@@ -37,7 +37,10 @@ void main() {
       final GoRouter router = await createGoRouter(tester);
       await tester.tap(find.text('Settings'));
       await tester.pumpAndSettle();
-      expect(router.location, '/page-0/settings?search=notification');
+      final ImperativeRouteMatch routeMatch = router
+          .routerDelegate.currentConfiguration.last as ImperativeRouteMatch;
+      expect(routeMatch.matches.uri.toString(),
+          '/page-0/settings?search=notification');
     });
   });
 }