+++ packages/pigeon/tool/run_tests.dart
@@ -38,8 +38,8 @@ Future<void> _validateGeneratedTestFiles() async {
   final String relativePigeonPath = p.relative(baseDir, from: repositoryRoot);
 
   print('Validating generated files:');
-  print('  Generating output...');
-  final int generateExitCode = await generatePigeons(baseDir: baseDir);
+  print('  Generating test output...');
+  final int generateExitCode = await generateTestPigeons(baseDir: baseDir);
   if (generateExitCode != 0) {
     print('Generation failed; see above for errors.');
     exit(generateExitCode);
@@ -62,6 +62,60 @@ Future<void> _validateGeneratedTestFiles() async {
   }
 
   print('The following files are not updated, or not formatted correctly:');
+  modifiedFiles.map((String line) => '  $line').forEach(print);
+
+  print('\nTo fix run "dart run tool/generate.dart --format" from the pigeon/ '
+      'directory, or apply the diff with the command below.\n');
+
+  final ProcessResult diffResult = await Process.run(
+    'git',
+    <String>['diff', relativePigeonPath],
+    workingDirectory: repositoryRoot,
+  );
+  if (diffResult.exitCode != 0) {
+    print('Unable to determine diff.');
+    exit(1);
+  }
+  print('patch -p1 <<DONE');
+  print(diffResult.stdout);
+  print('DONE');
+  exit(1);
+}
+
+Future<void> _validateGeneratedExampleFiles() async {
+  final String baseDir = p.dirname(p.dirname(Platform.script.toFilePath()));
+  final String repositoryRoot = p.dirname(p.dirname(baseDir));
+  final String relativePigeonPath = p.relative(baseDir, from: repositoryRoot);
+
+  print('Validating generated files:');
+  print('  Generating example output...');
+
+  final int generateExitCode = await generateExamplePigeons();
+
+  if (generateExitCode != 0) {
+    print('Generation failed; see above for errors.');
+    exit(generateExitCode);
+  }
+
+  print('  Formatting output...');
+  final int formatExitCode =
+      await formatAllFiles(repositoryRoot: repositoryRoot);
+  if (formatExitCode != 0) {
+    print('Formatting failed; see above for errors.');
+    exit(formatExitCode);
+  }
+
+  print('  Checking for changes...');
+  final List<String> modifiedFiles = await _modifiedFiles(
+      repositoryRoot: repositoryRoot, relativePigeonPath: relativePigeonPath);
+
+  if (modifiedFiles.isEmpty) {
+    return;
+  }
+
+  print(
+      'Either messages.dart and messages_test.dart have non-matching definitions or');
+  print('the following files are not updated, or not formatted correctly:');
   modifiedFiles.map((String line) => '  $line').forEach(print);
 
   print('\nTo fix run "dart run tool/generate.dart --format" from the pigeon/ '
@@ -121,13 +175,14 @@ Future<void> main(List<String> args) async {
   ];
   const List<String> macOSHostTests = <String>[
     iOSObjCUnitTests,
-    iOSObjCIntegrationTests,
     // Currently these are testing exactly the same thing as
-    // macOSSwiftIntegrationTests, so we don't need to run both by default. This
+    // macOS*IntegrationTests, so we don't need to run both by default. This
     // should be enabled if any iOS-only tests are added (e.g., for a feature
     // not supported by macOS).
+    // iOSObjCIntegrationTests,
     // iOSSwiftIntegrationTests,
     iOSSwiftUnitTests,
+    macOSObjCIntegrationTests,
     macOSSwiftUnitTests,
     macOSSwiftIntegrationTests,
   ];
@@ -164,6 +219,7 @@ Future<void> main(List<String> args) async {
       print('Skipping generated file validation on stable.');
     } else {
       await _validateGeneratedTestFiles();
+      await _validateGeneratedExampleFiles();
     }
   }