+++ script/tool/lib/src/gradle_check_command.dart
@@ -18,10 +18,13 @@ final Version minKotlinVersion = Version(1, 7, 10);
 /// A command to enforce gradle file conventions and best practices.
 class GradleCheckCommand extends PackageLoopingCommand {
   /// Creates an instance of the gradle check command.
-  GradleCheckCommand(Directory packagesDir) : super(packagesDir);
+  GradleCheckCommand(super.packagesDir);
 
   @override
   final String name = 'gradle-check';
+
+  @override
+  List<String> get aliases => <String>['check-gradle'];
 
   @override
   final String description =